Cleanup server.js
